@@ -1,31 +1,52 @@
 {
-  "name": "types-from-odata-v2",
-  "version": "1.0.0",
-  "description": "",
+  "name": "odata-v2-types-generator",
+  "license": "MIT",
+  "version": "0.0.2",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "exports": {
+    ".": {
+      "require": "./dist/cjs/index.js",
+      "import": "./dist/esm/index.js"
+    }
+  },
+  "type": "module",
   "scripts": {
     "test": "jest --verbose",
-    "build": "tsc",
+    "build": "rollup -c rollup.config.js",
     "dev": "nodemon --exec ts-node src/index.ts",
-    "start": "ts-node src/index.ts"
+    "start": "ts-node src/index.ts",
+    "format": "prettier . --write"
   },
-  "author": "",
-  "license": "ISC",
+  "author": "Oleg Borodatov",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/borolgs/odata-v2-types-generator"
+  },
   "dependencies": {
-    "fast-xml-parser": "^4.0.9",
-    "typescript": "^4.7.4"
+    "fast-xml-parser": "4.1.3",
+    "typescript": "4.9.5"
   },
   "devDependencies": {
-    "@types/jest": "^27.5.2",
-    "@types/node": "^17.0.45",
-    "@typescript-eslint/eslint-plugin": "^5.34.0",
-    "@typescript-eslint/parser": "^5.34.0",
-    "eslint": "^8.22.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^27.5.1",
-    "nodemon": "^2.0.19",
-    "prettier": "^2.7.1",
-    "ts-jest": "^27.1.5",
-    "ts-node": "^10.9.1"
+    "@rollup/plugin-commonjs": "^24.0.1",
+    "@types/jest": "27.5.2",
+    "@types/node": "17.0.45",
+    "@typescript-eslint/eslint-plugin": "5.55.0",
+    "@typescript-eslint/parser": "5.55.0",
+    "eslint": "8.36.0",
+    "eslint-config-prettier": "8.7.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "jest": "27.5.1",
+    "nodemon": "2.0.21",
+    "prettier": "2.8.4",
+    "rollup": "^3.19.1",
+    "rollup-plugin-dts": "^5.3.0",
+    "rollup-plugin-typescript2": "^0.34.1",
+    "ts-jest": "27.1.5",
+    "ts-node": "10.9.1"
   }
 }
